# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Assist Tix API Documentation",
        "title": "Assist Tix Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/venues": {
            "get": {
                "description": "Create venue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "venue"
                ],
                "summary": "Create venue",
                "responses": {
                    "200": {
                        "description": "List venues",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VenueResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create venue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "venue"
                ],
                "summary": "Create venue",
                "parameters": [
                    {
                        "description": "Create venue request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateVenueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venue created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            }
        },
        "/venues/{venueId}": {
            "get": {
                "description": "Get venue By ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "venue"
                ],
                "summary": "Get venue By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Venue ID",
                        "name": "venueId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get venue by id",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VenueResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit venue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "venue"
                ],
                "summary": "Edit venue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Venue ID",
                        "name": "venueId",
                        "in": "path"
                    },
                    {
                        "description": "update venue request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateVenueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venue created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete venue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "venue"
                ],
                "summary": "Delete venue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Venue ID",
                        "name": "venueId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateVenueRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "name",
                "status",
                "type"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "INACTIVE",
                        "DISABLE"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "STADIUM",
                        "VENUE",
                        "HALL",
                        "OTHER"
                    ]
                }
            }
        },
...
        "dto.UpdateVenueRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "name",
                "status",
                "type"
            ],
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "INACTIVE",
                        "DISABLE"
                    ]
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "STADIUM",
                        "VENUE",
                        "HALL",
                        "OTHER"
                    ]
                }
            }
        },
        "dto.VenueResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "lib.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "lib.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": false
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}